Share extractJson between agent + verifier judges (fix truncated-response FAIL)
RCA: the truncation-recovery fix (#327) was applied only to verifier-judge.ts.
agent-judge.ts kept a duplicate of the OLD indexOf('{')/lastIndexOf('}')
extractJson, so the dual/agent judge still FAILed genuine verdicts when the ACP
agent's reply was truncated/brace-short. Observed: gemma4:26b dual-judge
throughput + MCP runs went red despite correct output — the MCP agent's own
verdict was "pass": true, but the harness "Failed to parse agent response".

The duplication is the root cause of the drift. Extract the string-aware
brace-walk recovery into cicd/tests/src/judge/extract-json.ts and have BOTH
judges import it, so a future fix can't land in one copy and miss the other.
